These are the files mari-lexc.txt (--> root.lexc) and mr.script (--> mhr-phon.xfscript)
They are written by Jorma Luutonen, and given to the Giellatekno/Divvun project under CC-BY conditions:
The files may be distributed, changed, and used for any purpose, as long as the original author is contributed.

These two files contain the whole morphophonology and morphology for a small sample of nouns and verbs.
The root.lexc file will now be split according to POS.