This is the .#. problem checkin.
Inspired by smn (Lene) we replaced .#. in the
twolc file with %^END: and generated K further
on to the lexicon ENDLEX situated in root.lexc.

The point behind this was to situate a symbol
(here: %^END) to the left of end-of-string.
Before this, things worked in twolc but not
in the resulting fst. Now it works in both.

This is a xfst-needed hack: Things worked in 
hfst also without this additional ENDLEX business.

Yaml for xfst down to 37, but there are still open
issues.