Fix #5 – unknowns were not getting analysis
The unknownword definition has become a bit more complicated since
last, but possibly more explicit. It used to be enough to have an
empty analysis for unknowns and hfst-tokenise would magically output
empty analyses as "theform ?" – now we explicitly add a (heavily
weighted) analysis with "theform ?" in pmscript. So instead of
pmscript having

    unknownform:0

It needs

    unknownform 0:% %?%?::280284

ie. copy the unknownform to analysis and then add a space (% ) and
question mark (why twice though?) with weight 280284.